requirements.txt
# ══════════════════════════════════════════════════════════
# Harun AI Bot — requirements.txt
# Olusturma: pip freeze ciktisi filtrelenerek hazırlandi
# Kurulum: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════

# --- ZORUNLU: Bot calisması icin gerekli ---
python-telegram-bot==22.6
python-dotenv==1.2.1
groq==1.0.0
httpx==0.28.1
requests==2.32.5
beautifulsoup4==4.14.3
ddgs==9.10.0
python-docx==1.2.0
python-pptx==1.0.2
psutil==7.2.2
colorama==0.4.6
Flask==3.1.2
APScheduler==3.11.2

# --- OPSIYONEL: PC Kontrol / Otomasyon ---
PyAutoGUI==0.9.54
mss==10.1.0
pillow==12.0.0

# --- OPSIYONEL: Ses ---
SpeechRecognition==3.14.5
pyttsx3==2.99
pydub==0.25.1
PyAudio==0.2.14

# --- OPSIYONEL: RAG / Hafiza / Embedding ---
sentence-transformers==5.2.3
chromadb==1.5.0
numpy==2.4.2

# --- OPSIYONEL: Fine-tuning / ML ---
torch==2.10.0
transformers==4.57.6
datasets==4.3.0
trl==0.24.0
peft==0.18.1
accelerate==1.12.0
unsloth==2026.2.1
unsloth_zoo==2026.2.1
safetensors==0.7.0
bitsandbytes==0.49.2

# --- OPSIYONEL: PDF isleme ---
pdfminer.six==20251230
pdfplumber==0.11.9

# --- OPSIYONEL: Diger ---
ollama==0.6.1
openai==1.83.0
openpyxl==3.1.5
lxml==6.0.2
pandas==3.0.1
PyYAML==6.0.3
